moving .h to a proper directory
